adding paging to the things data

+++ Core/src/com/gawkat/core/client/account/thingstufftype/ThingStuffTypesData.java
@@ -4,7 +4,7 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class ThingStuffTypesData implements IsSerializable {
   
-  public int total = 0;
+  public long total = 0;
   
   public ThingStuffTypeData[] thingStuffTypeData = null;
